back/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
pydantic==2.9.2
groq==0.11.0
httpx==0.27.2
beautifulsoup4==4.12.3
PyJWT==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.1
